move video test cases to a single file
